Move JsonWebTokenTest and JsonWebTokenFactory to correct namespace
