+++ tests/Jwt/JsonWebTokenTest.php
@@ -2,9 +2,8 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Jwt\Symfony\Authentication;
+namespace CultuurNet\UDB3\Jwt;
 
-use CultuurNet\UDB3\Jwt\JsonWebToken;
 use CultuurNet\UDB3\User\UserIdentityDetails;
 use CultuurNet\UDB3\User\UserIdentityResolver;
 use PHPUnit\Framework\TestCase;
